chore: refactor-oc-table-row-to-setup-script
